test: 테스트 코드 수정
- 보너스 번호 입력에 대한 테스트 코스 유지보수

+++ src/test/java/lotto/domain/LottoCalculatorTest.java
@@ -23,7 +23,7 @@ class LottoCalculatorTest {
         // when
         LottoCalculator calculator = new LottoCalculator();
 
-        List<Integer> matchingCount = calculator.calculate(lottoTickets, winningLottoTicket);
+        List<Double> matchingCount = calculator.calculate(lottoTickets, winningLottoTicket);
 
         int count = calculator.getCountOfWin(matchingCount, 4);
 
@@ -44,7 +44,7 @@ class LottoCalculatorTest {
 
         // when
         LottoCalculator calculator = new LottoCalculator();
-        List<Integer> matchingCount = calculator.calculate(lottoTickets, winningLottoTicket);
+        List<Double> matchingCount = calculator.calculate(lottoTickets, winningLottoTicket);
 
         int prizeMoney = calculator.getPrizeMoney(matchingCount);
